ext2fuse: Divide third_party.fuse.ext2fuse.module
Now it's fs driver and cmd
