+++ src/cmds/fs/ext2fuse/Mybuild
@@ -1,0 +1,25 @@
+package embox.cmd.fs
+
+@AutoCmd
+@Cmd(name = "ext2fuse",
+	help = "Ext2fuse module",
+	man = '''
+		NAME
+			ext2fuse - start file "ext2" system in userspace
+		SYNOPSIS
+			ext2fuse bdev_name mount_point
+		DESCRIPTION
+		OPTIONS
+		AUTHORS
+			Anton Bondarev
+	''')
+module ext2fuse {
+	@IncludePath("$(EXTERNAL_BUILD_DIR)/third_party/fuse/core/install/include")
+	@IncludePath("$(ROOT_DIR)/third-party/fuse/ext2fuse/include")
+	@IncludePath("$(ROOT_DIR)/third-party/fuse/fuse")
+	source "ext2fuse.c"
+
+	@NoRuntime depends third_party.ext2fuse.core
+	depends embox.fs.driver.ext2fuse
+	depends embox.fs.fuse.fuse_linux
+}
