+++ src/fs/driver/ext2fuse/Mybuild
@@ -1,0 +1,11 @@
+package embox.fs.driver
+
+module ext2fuse {
+	@IncludePath("$(EXTERNAL_BUILD_DIR)/third_party/fuse/core/install/include")
+	@IncludePath("$(ROOT_DIR)/third-party/fuse/ext2fuse/include")
+	@IncludePath("$(ROOT_DIR)/third-party/fuse/fuse")
+	source "ext2fuse.c"
+
+	@NoRuntime depends third_party.ext2fuse.core
+	depends embox.fs.fuse.fuse_linux
+}
